V0.0.2 (2025-08-28)
Version 0.0.2 enhances the library by introducing a builder-oriented profile construction and adaptation API, as well as additional tag types for comprehensive high-level modifications to tag values.

Furthermore, a substantial number of tests have been incorporated, including a round-trip test for both reading and writing ICC profiles, as well as a test for the binary equality of several standard color profiles.